pyproject.toml
[project]
name = "mail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.155.0",
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.1",
    "httpx>=0.28.1",
    "mcp>=1.1.2",
    "pyautogui>=0.9.54",
    "pyperclip>=1.9.0",
]
[build-system]
requires = [ "hatchling" ,]
build-backend = "hatchling.build"
[project.scripts]
mail = "mail:main"

